requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-ninja==0.19.1
h11==0.14.0
pydantic==1.10.2
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uvicorn==0.20.0
